+++ api/adm_groups_pagination.php
@@ -21,7 +21,7 @@
 
  if($jwt){
     try {
-
+        $decoded = JWT::decode($jwt, $key, array('HS256'));
         $limit = 5;  
         if (isset($_GET["page"])) { $page  = $_GET["page"]; } else { $page=1; };  
         $start_from = ($page-1) * $limit;